Fix issue with a message separating line containing a space.
The DBC file contains two messages that are separated by a line with one space.
In previous versions this wasn't interpreted as message separator.
Hence the signals of the second message were appended to the first message.
chomp was modified to also chop away spaces and tabs.
A regression test was added.

Signed-off-by: Tobias Lorenz <tobias.lorenz@gmx.net>