Allow to read files that have messages without blank line as separator
This is actually against the standard, but is seen in the wild.
Now reading of signals doesn't happen in a loop within the message read
any more. Instead the last message context is remembered and the signal
is parsed directly by the main loop.

A new Status::SignalWithoutMessage was introduced to indicate that an
SG line was read, without a preceding BO or SG line.

Signed-off-by: Tobias Lorenz <tobias.lorenz@gmx.net>